wayland: move input logic to the event loop
